# web/dashboard/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing LoRaWAN rain barrel devices and sensor data",
        "title": "Regenfass API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:64000",
    "basePath": "/",
    "paths": {
        "/device": {
            "post": {
                "description": "Register a device using it's device EUI",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Register a device",
                "parameters": [
                    {
                        "description": "Device to register",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RegisterDevicePayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.RegisterDeviceResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Device EUI is already registered",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
        "/device/{deviceToken}": {
            "get": {
                "description": "Retrieve device details using either read-write or read-only token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Get device information by token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device token (read-write or read-only)",
                        "name": "deviceToken",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.DeviceInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Device not found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update mutable device fields (currently only the custom name). Requires the read-write token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Update device fields",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device read-write token",
                        "name": "deviceToken",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateDevicePayload"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid token or payload",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Device token is read-only",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Device not found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
        "/device/{deviceToken}/channels/{channelID}": {
            "put": {
                "description": "Set the name and declared measurement type of one of a device's channels. The channel does not have to have reported anything yet — describing it up front is how a slot is prepared for a sensor. Requires the read-write token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Describe a device channel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device read-write token",
                        "name": "deviceToken",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Channel id (0–15)",
                        "name": "channelID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "How to describe the channel",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpsertDeviceChannelPayload"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid token, channel id or payload",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Device token is read-only",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Device not found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
        "/device/{deviceToken}/channels/{channelID}/hidden": {
            "put": {
                "description": "Takes a channel off the dashboard, or puts it back. Nothing is deleted: the channel's measurements stay and new ones keep arriving, they are simply not returned by the measurement endpoints while it is hidden. A hidden channel still appears in a device's `channels` so it can be restored. Requires the read-write token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Hide or restore a device channel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device read-write token",
                        "name": "deviceToken",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Channel id (0–15)",
                        "name": "channelID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Whether to hide the channel",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SetDeviceChannelHiddenPayload"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid token, channel id or payload",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Device token is read-only",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Device not found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.DeviceInfoResponse": {
            "description": "Identity of a device. When authenticated with the read-write token, both tokens are returned; otherwise only the read-only token is returned. `name` is the user-set name, or a stable auto-generated nickname if none has been set.",
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "is_readonly": {
                    "type": "boolean",
                    "example": false
                },
                "latitude": {
                    "type": "number",
                    "example": 51.3397
                },
                "longitude": {
                    "type": "number",
                    "example": 12.3731
                },
                "name": {
                    "type": "string",
                    "example": "Happy Barrel"
                },
                "read_only_token": {
                    "type": "string",
                    "example": "ro_token_123"
                },
                "read_write_token": {
                    "type": "string",
                    "example": "rw_token_123"
                }
            }
        },
...
        "api.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "api.RegisterDevicePayload": {
            "type": "object",
            "properties": {
                "device_eui": {
                    "type": "string",
                    "example": "AABBCCDDEEFF0011"
                }
            }
        },
        "api.RegisterDeviceResponse": {
            "type": "object",
            "properties": {
                "read_only_token": {
                    "type": "string",
                    "example": "ro_token_123"
                },
                "read_write_token": {
                    "type": "string",
                    "example": "rw_token_123"
                }
            }
        },
        "api.SetDeviceChannelHiddenPayload": {
            "description": "Whether the channel should be shown on the dashboard.",
            "type": "object",
            "properties": {
                "hidden": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "api.UpdateDevicePayload": {
            "description": "Fields that can be updated on a device (only present fields are applied)",
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Rain barrel by the shed"
                }
            }
        },
        "api.UpsertDeviceChannelPayload": {
            "description": "How a device's channel should be described. `measurement_type` is the type the user picked in the dashboard. Send either field as null (or the name as empty) to clear it, which returns the channel to being undescribed.",
            "type": "object",
            "properties": {
                "measurement_type": {
                    "type": "integer",
                    "example": 4
                },
                "name": {
                    "type": "string",
                    "example": "Water Level"
                }
            }
        }
    }
}